+++ htrest/apis/time_prog.py
@@ -24,12 +24,12 @@ from flask import request
 from flask_restx import Namespace, Resource, fields
 from htheatpump.httimeprog import TimeProgram as HtTimeProg
 from htheatpump.httimeprog import TimeProgEntry as HtTimeProgEntry
+from ..app import ht_heatpump
+from .. import settings
 from .utils import HtContext
-from htrest.app import ht_heatpump
-from htrest import settings
 
 
-_logger = logging.getLogger(__name__)
+_LOGGER = logging.getLogger(__name__)
 
 api = Namespace(
     "timeprog", description="Operations related to the time programs of the heat pump."
@@ -123,7 +123,7 @@ class TimeProgs(Resource):
     @api.marshal_list_with(time_prog_model, skip_none=True)
     def get(self):
         """ Returns a list of all available time programs of the heat pump. """
-        _logger.info("*** [GET] {!s}".format(request.url))
+        _LOGGER.info("*** [GET] %s", request.url)
         with HtContext(ht_heatpump):
             time_progs = ht_heatpump.get_time_progs()
         return [time_prog.as_json(with_entries=False) for time_prog in time_progs]
@@ -135,7 +135,7 @@ class TimeProg(Resource):
     @api.marshal_with(time_prog_with_entries_model, skip_none=True)
     def get(self, id: int):
         """ Returns the time program with the given index of the heat pump. """
-        _logger.info("*** [GET] {!s} -- id={}".format(request.url, id))
+        _LOGGER.info("*** [GET] %s -- id=%d", request.url, id)
         with HtContext(ht_heatpump):
             time_prog = ht_heatpump.get_time_prog(id)
         return time_prog.as_json(with_entries=True)
@@ -144,13 +144,12 @@ class TimeProg(Resource):
     @api.marshal_with(time_prog_with_entries_model)
     def put(self, id: int):
         """ Sets all time program entries of a specific time program of the heat pump. """
-        _logger.info(
-            "*** [PUT{}] {!s} -- id={}, payload={!s}".format(
-                " (read-only)" if settings.READ_ONLY else "",
-                request.url,
-                id,
-                api.payload,
-            )
+        _LOGGER.info(
+            "*** [PUT%s] %s -- id=%d, payload=%s",
+            " (read-only)" if settings.READ_ONLY else "",
+            request.url,
+            id,
+            api.payload,
         )
         with HtContext(ht_heatpump):
             time_prog = ht_heatpump.get_time_prog(id, with_entries=False).as_json(
@@ -173,9 +172,7 @@ class TimeProgEntry(Resource):
     @api.marshal_with(time_prog_entry_model)
     def get(self, id: int, day: int, num: int):
         """ Returns a specific time program entry of the heat pump. """
-        _logger.info(
-            "*** [GET] {!s} -- id={}, day={}, num={}".format(request.url, id, day, num)
-        )
+        _LOGGER.info("*** [GET] %s -- id=%d, day=%d, num=%d", request.url, id, day, num)
         with HtContext(ht_heatpump):
             entry = ht_heatpump.get_time_prog_entry(id, day, num)
         return entry.as_json()
@@ -184,15 +181,14 @@ class TimeProgEntry(Resource):
     @api.marshal_with(time_prog_entry_model)
     def put(self, id: int, day: int, num: int):
         """ Sets a specific time program entry of the heat pump. """
-        _logger.info(
-            "*** [PUT{}] {!s} -- id={}, day={}, num={}, payload={!s}".format(
-                " (read-only)" if settings.READ_ONLY else "",
-                request.url,
-                id,
-                day,
-                num,
-                api.payload,
-            )
+        _LOGGER.info(
+            "*** [PUT%s] %s -- id=%d, day=%d, num=%d, payload=%s",
+            " (read-only)" if settings.READ_ONLY else "",
+            request.url,
+            id,
+            day,
+            num,
+            api.payload,
         )
         entry = HtTimeProgEntry.from_json(api.payload)
         with HtContext(ht_heatpump):